% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutByShp.R
\name{cutByShp}
\alias{cutByShp}
\title{Cut A Data Frame by A Shapefile}
\usage{
cutByShp(myshp, dat, lat.name, long.name)
}
\arguments{
\item{myshp}{the shapefile (class SpatialPolygonsDataFrame)}

\item{dat}{the data frame to be cut}
}
\value{
A subset of the data frame cut by the shapefile
}
\description{
Cut a data frame by a shapefile
}
\examples{
cutByShp(myshp = shp, dat = df, lat.name = 'Lat', long.name = 'Lon')
}
