% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPurpleairLst.R
\name{getPurpleairLst}
\alias{getPurpleairLst}
\title{Get a list of PurpleAir sensors}
\usage{
getPurpleairLst(output.path)
}
\arguments{
\item{output.path}{the path of the output CSV file.}
}
\description{
Get a list of the latest PurpleAir sensors from PurpleAir JSON (\link{https://www.purpleair.com/json}) and save as a CSV file.
}
\examples{
getPurpleairLst('/path')
}
