% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movAvgLag.R
\name{movAvgLag}
\alias{movAvgLag}
\title{Moving Average Creator}
\usage{
movAvgLag(data, lag1, lag2)
}
\arguments{
\item{data}{a vector for which the moving average is created}

\item{lag1}{a non-negative integer indicating the starting lag day (lag2 > lag1)}

\item{lag2}{a non-negative integer indicating the ending lag day (lag2 > lag1)}
}
\value{
A `crossbasis` matrix showing the moving average
}
\description{
Moving Average Creator
}
\examples{
MA 0-3
movAvgLag(data = df$pm25, lag1 = 0, lag2 = 3)
}
