% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purpleairDownload.R
\name{purpleairDownload}
\alias{purpleairDownload}
\title{Download Purple Air PM2.5 data}
\usage{
purpleairDownload(site.csv, start.date, end.date, output.path, average,
  time.zone = "GMT", indoor = F)
}
\arguments{
\item{site.csv}{a data frame of a site list or a absolute path to the site CSV file (from \code{getPurpleairLst}).}

\item{start.date}{the beginning date in the format "YYYY-MM-DD".}

\item{end.date}{the end date in the format "YYYY-MM-DD".}

\item{output.path}{the path to output CSV files.}

\item{average}{get average of this many minutes, valid values: 10, 15, 20, 30, 60, 240, 720, 1440, "daily". "daily" is not recommended as the daily values can only be calculated at the UTC time.}

\item{time.zone}{time zone specification to be used for the conversion, but "" is the current time zone, and "GMT" is UTC (Universal Time, Coordinated). Invalid values are most commonly treated as UTC, on some platforms with a warning.}

\item{indoor}{whether includes indoor sites (FALSE by default).}
}
\description{
Download Purple Air PM2.5 data and save as csv files (each PurpleAir site per file). The indoor sites are not included by default.
}
\examples{
purpleairDownload(site.csv = '/absolute/path/to/the/sensorlist.csv',
    start.date = '2017-01-01',
    end.date = '2017-12-31',
    output.path = '/output_path',
    average = 60,
    time.zone = 'America/Los_Angeles')
}
