% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QA2Char.R
\name{QA2Char}
\alias{QA2Char}
\title{Convert Decimal to MAIAC Quality Flags}
\usage{
QA2Char(num)
}
\arguments{
\item{num}{a decimal number less than or equal to 2^16 (since the QA flag only has 16 digits)}
}
\value{
An array of strings regarding the MAIAC QA flags
}
\description{
Convert a decimal number to MAIAC quality flags
}
\examples{
QA2Char(8)
}
