% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unconDistLag.R
\name{unconDistLag}
\alias{unconDistLag}
\title{Unconstrained Distributed Lag Creator}
\usage{
unconDistLag(data, lag1, lag2)
}
\arguments{
\item{data}{a vector for which the distributed lag is created}

\item{lag1}{a non-negative integer indicating the starting lag day (lag2 > lag1)}

\item{lag2}{a non-negative integer indicating the ending lag day (lag2 > lag1)}
}
\value{
A `crossbasis` matrix showing the distributed lag
}
\description{
Unconstrained Distributed Lag Creator
}
\examples{
# Lag 0-3
unconDistLag(data = df$pm25, lag1 = 0, lag2 = 3)
}
