% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMCD19A2.R
\name{readMCD19A2}
\alias{readMCD19A2}
\title{Read MCD19A2 HDF-4 Files}
\usage{
readMCD19A2(file.name, latlong.range = NULL, border.shp = NULL)
}
\arguments{
\item{file.name}{the name of input HDF file with an \emph{absolute} path}

\item{latlong.range}{the lat/long range with the structure \code{c(lon.min, lon.max, lat.min, lat.max)}, cannot be used with \code{border.shp}}

\item{border.shp}{a shapefile (class SpatialPolygonsDataFrame), cannot be used with \code{latlong.range}}
}
\value{
A list with time stamps and MAIAC AOD data in multiple bands
}
\description{
Read MCD19A2 HDF-4 files
}
\examples{
readMCD19A2(file.name = '/path/file.HDF', latlong.range = c(132, 133, 56, 57))
readMCD19A2(file.name = '/path/file.HDF', border.shp = myshp)
}
